+++ boards/stm32h7/src/macros.rs
@@ -1,4 +1,5 @@
-macro_rules! config_matrix_pins {
+// DEPRECIATED
+macro_rules! _config_matrix_pins {
     (input: [$($in_port:ident.$in_pin:ident), *], output: [$($out_port:ident.$out_pin:ident), +]) => {
         {
             $(
@@ -17,20 +18,14 @@ macro_rules! config_matrix_pins {
     };
 }
 
-macro_rules! config_matrix_pins_embassy {
+macro_rules! config_matrix_pins_stm32 {
     (peripherals: $p:ident, input: [$($in_pin:ident), *], output: [$($out_pin:ident), +]) => {
         {
-            $(
-                let $in_pin = Input::new($p.$in_pin, embassy_stm32::gpio::Pull::Down).degrade();
-            )*
-            $(
-                let mut $out_pin = Output::new($p.$out_pin, embassy_stm32::gpio::Level::Low, embassy_stm32::gpio::Speed::VeryHigh).degrade();
-            )+
-            $(
-                $out_pin.set_low();
-            )+
-            let output_pins = [$($out_pin), +];
-            let input_pins = [$($in_pin), +];
+            let mut output_pins = [$(Output::new($p.$out_pin, embassy_stm32::gpio::Level::Low, embassy_stm32::gpio::Speed::VeryHigh).degrade()), +];
+            let input_pins = [$(Input::new($p.$in_pin, embassy_stm32::gpio::Pull::Down).degrade()), +];
+            output_pins.iter_mut().for_each(|p| {
+                p.set_low();
+            });
             (input_pins, output_pins)
         }
     };
