+++ rmk/src/lib.rs
@@ -7,15 +7,13 @@
 #![cfg_attr(not(test), no_std)]
 
 use action::KeyAction;
-use config::KeyboardConfig;
 use core::convert::Infallible;
 use eeprom::{eeconfig::Eeconfig, EepromStorageConfig};
 use embedded_hal::digital::{InputPin, OutputPin};
 use embedded_storage::nor_flash::NorFlash;
 use keyboard::Keyboard;
 use usb::KeyboardUsbDevice;
-use usb2::KeyboardUsbDevice2;
-use usb_device::class_prelude::{UsbBus, UsbBusAllocator};
+use embassy_usb::{class::hid::State, driver::Driver};
 
 pub use embassy_sync;
 pub use embassy_usb;
@@ -33,47 +31,46 @@ pub mod keymap;
 pub mod layout_macro;
 pub mod matrix;
 pub mod usb;
-pub mod usb2;
 pub mod via;
 
-/// Initialize keyboard core and keyboard usb device
-pub fn initialize_keyboard_and_usb_device<
-    'a,
-    B: UsbBus,
-    In: InputPin<Error = Infallible>,
-    Out: OutputPin<Error = Infallible>,
-    F: NorFlash,
-    const EEPROM_SIZE: usize,
-    const ROW: usize,
-    const COL: usize,
-    const NUM_LAYER: usize,
->(
-    usb_allocator: &'a UsbBusAllocator<B>,
-    config: &KeyboardConfig<'a>,
-    storage: Option<F>,
-    eeprom_storage_config: EepromStorageConfig,
-    eeconfig: Option<Eeconfig>,
-    input_pins: [In; ROW],
-    output_pins: [Out; COL],
-    keymap: [[[KeyAction; COL]; ROW]; NUM_LAYER],
-) -> (
-    Keyboard<In, Out, F, EEPROM_SIZE, ROW, COL, NUM_LAYER>,
-    KeyboardUsbDevice<'a, B>,
-) {
-    (
-        Keyboard::new(
-            input_pins,
-            output_pins,
-            storage,
-            eeprom_storage_config,
-            eeconfig,
-            keymap,
-        ),
-        KeyboardUsbDevice::new(usb_allocator, config),
-    )
-}
-use embassy_usb::{class::hid::State, driver::Driver};
+// DEPRECIATED
+// pub fn initialize_keyboard_and_usb_device<
+//     'a,
+//     B: UsbBus,
+//     In: InputPin<Error = Infallible>,
+//     Out: OutputPin<Error = Infallible>,
+//     F: NorFlash,
+//     const EEPROM_SIZE: usize,
+//     const ROW: usize,
+//     const COL: usize,
+//     const NUM_LAYER: usize,
+// >(
+//     usb_allocator: &'a UsbBusAllocator<B>,
+//     config: &KeyboardConfig<'a>,
+//     storage: Option<F>,
+//     eeprom_storage_config: EepromStorageConfig,
+//     eeconfig: Option<Eeconfig>,
+//     input_pins: [In; ROW],
+//     output_pins: [Out; COL],
+//     keymap: [[[KeyAction; COL]; ROW]; NUM_LAYER],
+// ) -> (
+//     Keyboard<In, Out, F, EEPROM_SIZE, ROW, COL, NUM_LAYER>,
+//     KeyboardUsbDevice<'a, B>,
+// ) {
+//     (
+//         Keyboard::new(
+//             input_pins,
+//             output_pins,
+//             storage,
+//             eeprom_storage_config,
+//             eeconfig,
+//             keymap,
+//         ),
+//         KeyboardUsbDevice::new(usb_allocator, config),
+//     )
+// }
 
+/// Initialize keyboard core and keyboard usb device
 pub fn initialize_keyboard_and_usb_device2<
     'd,
     D: Driver<'d>,
@@ -95,7 +92,7 @@ pub fn initialize_keyboard_and_usb_device2<
     keymap: [[[KeyAction; COL]; ROW]; NUM_LAYER],
 ) -> (
     Keyboard<In, Out, F, EEPROM_SIZE, ROW, COL, NUM_LAYER>,
-    KeyboardUsbDevice2<'d, D>,
+    KeyboardUsbDevice<'d, D>,
 ) {
     (
         Keyboard::new(
@@ -106,6 +103,6 @@ pub fn initialize_keyboard_and_usb_device2<
             eeconfig,
             keymap,
         ),
-        KeyboardUsbDevice2::new(driver, state),
+        KeyboardUsbDevice::new(driver, state),
     )
 }
